webapp/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
anyio==4.6.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
apscheduler==3.10.4
    # via -r requirements.in
beautifulsoup4==4.12.3
    # via python-fasthtml
bidict==0.23.1
    # via python-socketio
blinker==1.8.2
    # via flask
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   flask
    #   uvicorn
exceptiongroup==1.2.2
    # via anyio
fastcore==1.7.11
    # via
    #   fastlite
    #   python-fasthtml
    #   sqlite-minutils
fastlite==0.0.11
    # via python-fasthtml
flask==3.0.3
    # via
    #   -r requirements.in
    #   flask-socketio
flask-socketio==5.3.6
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
    #   wsproto
httpcore==1.0.6
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.2
    # via python-fasthtml
idna==3.10
    # via
    #   anyio
    #   httpx
itsdangerous==2.2.0
    # via
    #   flask
    #   python-fasthtml
jetson-gpio==2.1.6
    # via -r requirements.in
jinja2==3.1.4
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
numpy==2.1.1
    # via
    #   -r requirements.in
    #   tifffile
oauthlib==3.2.2
    # via python-fasthtml
packaging==24.1
    # via fastcore
pyserial==3.5
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via python-fasthtml
python-dotenv==1.0.1
    # via uvicorn
python-engineio==4.9.1
    # via python-socketio
python-fasthtml==0.2.1
    # via -r requirements.in
python-multipart==0.0.12
    # via python-fasthtml
python-socketio==5.11.4
    # via flask-socketio
pytz==2024.2
    # via apscheduler
pyyaml==6.0.2
    # via uvicorn
simple-websocket==1.0.0
    # via python-engineio
six==1.16.0
    # via
    #   apscheduler
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
soupsieve==2.6
    # via beautifulsoup4
sqlite-minutils==3.37.0.post3
    # via fastlite
starlette==0.39.2
    # via python-fasthtml
tifffile==2024.7.24
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   anyio
    #   uvicorn
tzlocal==5.2
    # via apscheduler
uvicorn[standard]==0.31.0
    # via python-fasthtml
uvloop==0.20.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
werkzeug==3.0.4
    # via flask
wsproto==1.2.0
    # via simple-websocket
